Clean up references from ember compatibility functions and make the data model implementation of reporting and attribute path expand iterator be the only implemented versions
