+++ src/app/BUILD.gn
@@ -199,8 +199,8 @@ static_library("interaction-model") {
     #
     #       This breaks having `.h` and `.cpp` based off the same compilation unit and
     #       should ideally be cleaned up.
-    "reporting/Read-DataModel.cpp",
-    "reporting/Read-DataModel.h",
+    "reporting/Read.cpp",
+    "reporting/Read.h",
     "reporting/reporting.h",
   ]
 
@@ -216,11 +216,11 @@ static_library("interaction-model") {
     ":constants",
     ":paths",
     ":subscription-info-provider",
+    "${chip_root}/src/app/MessageDef",
     "${chip_root}/src/app/data-model-provider",
     "${chip_root}/src/app/icd/server:icd-server-config",
     "${chip_root}/src/app/icd/server:manager",
     "${chip_root}/src/app/icd/server:observer",
-    "${chip_root}/src/app/MessageDef",
     "${chip_root}/src/app/util:af-types",
     "${chip_root}/src/app/util:callbacks",
     "${chip_root}/src/lib/address_resolve",
@@ -397,6 +397,8 @@ static_library("app") {
   output_name = "libCHIPDataModel"
 
   sources = [
+    "AttributePathExpandIterator.cpp",
+    "AttributePathExpandIterator.h",
     "AttributePathExpandIterator.h",
     "AttributePersistenceProvider.h",
     "ChunkedWriteCallback.cpp",
@@ -436,6 +438,7 @@ static_library("app") {
     ":global-attributes",
     ":interaction-model",
     "${chip_root}/src/app/data-model",
+    "${chip_root}/src/app/data-model-provider",
     "${chip_root}/src/app/icd/server:icd-server-config",
     "${chip_root}/src/app/util:callbacks",
     "${chip_root}/src/lib/address_resolve",
@@ -444,32 +447,6 @@ static_library("app") {
     "${chip_root}/src/protocols/interaction_model",
     "${chip_root}/src/system",
   ]
-
-  if (chip_use_data_model_interface == "disabled") {
-    sources += [
-      "AttributePathExpandIterator-Ember.cpp",
-      "AttributePathExpandIterator-Ember.h",
-      "AttributePathExpandIterator.h",
-    ]
-  } else if (chip_use_data_model_interface == "check") {
-    sources += [
-      "AttributePathExpandIterator-Checked.cpp",
-      "AttributePathExpandIterator-Checked.h",
-      "AttributePathExpandIterator-DataModel.cpp",
-      "AttributePathExpandIterator-DataModel.h",
-      "AttributePathExpandIterator-Ember.cpp",
-      "AttributePathExpandIterator-Ember.h",
-      "AttributePathExpandIterator.h",
-    ]
-    public_deps += [ "${chip_root}/src/app/data-model-provider" ]
-  } else {  # enabled
-    sources += [
-      "AttributePathExpandIterator-DataModel.cpp",
-      "AttributePathExpandIterator-DataModel.h",
-      "AttributePathExpandIterator.h",
-    ]
-    public_deps += [ "${chip_root}/src/app/data-model-provider" ]
-  }
 
   if (chip_enable_read_client) {
     sources += [
