+++ src/app/chip_data_model.cmake
@@ -70,15 +70,10 @@ endfunction()
 #
 function(chip_configure_data_model APP_TARGET)
     set(SCOPE PRIVATE)
-    set(ADD_EMBER_INTERFACE_FILES TRUE)
-    cmake_parse_arguments(ARG "SKIP_EMBER_INTERFACE" "SCOPE;ZAP_FILE;IDL" "EXTERNAL_CLUSTERS" ${ARGN})
+    cmake_parse_arguments(ARG "" "SCOPE;ZAP_FILE;IDL" "EXTERNAL_CLUSTERS" ${ARGN})
 
     if(ARG_SCOPE)
         set(SCOPE ${ARG_SCOPE})
-    endif()
-
-    if(ARG_SKIP_EMBER_INTERFACE)
-        set(ADD_EMBER_INTERFACE_FILES FALSE)
     endif()
 
     # CMAKE data model auto-includes the server side implementation
@@ -173,10 +168,4 @@ function(chip_configure_data_model APP_TARGET)
         ${APP_GEN_FILES}
         ${APP_TEMPLATES_GEN_FILES}
     )
-
-    if(ADD_EMBER_INTERFACE_FILES)
-        target_sources(${APP_TARGET} ${SCOPE}
-           ${CHIP_APP_BASE_DIR}/util/ember-compatibility-functions.cpp
-        )
-    endif()
 endfunction()
