+++ src/app/clusters/pump-configuration-and-control-server/pump-configuration-and-control-server.cpp
@@ -25,7 +25,6 @@
 #include <app/InteractionModelEngine.h>
 #include <app/util/attribute-storage.h>
 #include <app/util/config.h>
-#include <app/util/ember-compatibility-functions.h>
 
 using namespace chip;
 using namespace chip::app;
