+++ src/darwin/Framework/CHIP/ServerEndpoint/MTRServerAccessControl.mm
@@ -35,7 +35,6 @@
 #include <lib/core/Global.h>
 #include <lib/core/NodeId.h>
 
-#include <app/util/ember-compatibility-functions.h>
 #include <app/util/privilege-storage.h>
 
 using namespace chip;
